refactor(cell): move cell-group folder
