--- src/cell-group/cell-group.md
@@ -1,9 +1,0 @@
-:: BASE_DOC ::
-
-## API
-### CellGroup Props
-
-名称 | 类型 | 默认值 | 说明 | 必传
--- | -- | -- | -- | --
-bordered | Boolean | - | 是否显示组边框 | N
-title | String | - | 单元格组标题 | N
