--- src/cell-group/style/css.js
@@ -1,1 +1,0 @@
-import './index.css';
