--- src/cell-group/style/index.js
@@ -1,1 +1,0 @@
-import '../../_common/style/mobile/components/cell-group/_index.less';
