--- src/cell-group/type.ts
@@ -1,17 +1,0 @@
-/* eslint-disable */
-
-/**
- * 该文件为脚本自动生成文件，请勿随意修改。如需修改请联系 PMC
- * */
-
-export interface TdCellGroupProps {
-  /**
-   * 是否显示组边框
-   */
-  bordered?: boolean;
-  /**
-   * 单元格组标题
-   * @default ''
-   */
-  title?: string;
-}
