+++ src/cell/cell-group.vue
@@ -19,7 +19,7 @@
 <script lang="ts">
 import { defineComponent } from 'vue';
 import config from '../config';
-import CellGroupProps from './props';
+import CellGroupProps from './cell-group-props';
 
 const { prefix } = config;
 const name = `${prefix}-cell-group`;
