+++ src/components.ts
@@ -2,8 +2,7 @@ export { default as Button } from './button';
 export { default as ButtonGroup } from './button-group';
 export { default as Divider } from './divider';
 export { default as Calendar } from './calendar';
-export { default as Cell } from './cell';
-export { default as CellGroup } from './cell-group';
+export { default as Cell, CellGroup } from './cell';
 export { default as Checkbox, CheckboxGroup } from './checkbox';
 export { default as Input } from './input';
 export { default as Textarea } from './textarea';
